refactor(lab): split the notebook app into per-concern modules
Every source file was a grab bag: store.ts held signals, persistence, cell
mutations and output trimming; protocol.ts held both state and execution
commands; kernel-source.ts was a single 1000-line template string. Split
along the boundaries that were already implicit:

  kernel/     worker source (11 String.raw parts), lifecycle, bridge, store ops
  state/      signals, persistence, cell mutations, run orchestration
  lib/        pure helpers — chart config/render, media, markdown, trim
  components/ App shell, sidebar, cell rows, and the output views
  protocol/   state, run, notebook and export command maps
  styles/     one sheet per area, imported in cascade order by main.ts

No behavior change: dependencies now point one way (components -> state ->
kernel/lib), no barrel files, and main.ts still spreads plain command-map
literals so the manifest extractor can read them statically. The kernel
split is byte-exact — each part carries the newline that follows its
opening backtick — so kernel stack-trace line numbers are unchanged.

AGENTS.md documents the new layout and the conventions that keep it.
