+++ apps/lab/src/components/JsonView.ts
