+++ apps/lab/src/lib/data-url.ts
@@ -1,0 +1,25 @@
+/** Data-URL <-> blob plumbing. No app concepts here — pure browser conversion. */
+
+export function dataUrlToBlob(dataUrl: string): Blob {
+  const comma = dataUrl.indexOf(',');
+  const head = dataUrl.slice(0, comma);
+  const mime = /data:([^;,]+)/.exec(head)?.[1] || 'application/octet-stream';
+  const body = dataUrl.slice(comma + 1);
+  if (head.includes(';base64')) {
+    const bin = atob(body);
+    const bytes = new Uint8Array(bin.length);
+    for (let i = 0; i < bin.length; i++) bytes[i] = bin.charCodeAt(i);
+    return new Blob([bytes], { type: mime });
+  }
+  return new Blob([decodeURIComponent(body)], { type: mime });
+}
+
+/** Browser download of a data URL, for the in-app save button. */
+export function downloadDataUrl(dataUrl: string, filename: string): void {
+  const a = document.createElement('a');
+  a.href = dataUrl;
+  a.download = filename;
+  document.body.appendChild(a);
+  a.click();
+  a.remove();
+}
