+++ apps/lab/src/lib/media.ts
@@ -1,20 +1,7 @@
 import { storage } from '@bundled/yaar';
-import { chartToPNG } from './chart';
-import type { ChartSpec } from './types';
-
-export function dataUrlToBlob(dataUrl: string): Blob {
-  const comma = dataUrl.indexOf(',');
-  const head = dataUrl.slice(0, comma);
-  const mime = /data:([^;,]+)/.exec(head)?.[1] || 'application/octet-stream';
-  const body = dataUrl.slice(comma + 1);
-  if (head.includes(';base64')) {
-    const bin = atob(body);
-    const bytes = new Uint8Array(bin.length);
-    for (let i = 0; i < bin.length; i++) bytes[i] = bin.charCodeAt(i);
-    return new Blob([bytes], { type: mime });
-  }
-  return new Blob([decodeURIComponent(body)], { type: mime });
-}
+import { chartToPNG } from './chart-render';
+import { dataUrlToBlob } from './data-url';
+import type { ChartSpec } from '../types';
 
 /** Normalise a caller-supplied path into the shared media tree under media/lab/. */
 export function mediaPath(raw?: string, fallbackName?: string): string {
@@ -44,14 +31,4 @@ export async function saveChart(
 ): Promise<{ path: string; uri: string; bytes: number }> {
   const png = await chartToPNG(spec, opts);
   return await saveDataUrl(png, mediaPath(path));
-}
-
-/** Browser download of a data URL, for the in-app save button. */
-export function downloadDataUrl(dataUrl: string, filename: string): void {
-  const a = document.createElement('a');
-  a.href = dataUrl;
-  a.download = filename;
-  document.body.appendChild(a);
-  a.click();
-  a.remove();
 }
