+++ apps/lab/src/state/run.ts
@@ -1,7 +1,9 @@
 import { createSignal } from '@bundled/solid-js';
-import { runInKernel, DEFAULT_TIMEOUT_MS } from './kernel';
-import { current, setCellOutput, setRunningCell, findCell } from './store';
-import type { CellOutput, OutputPart, RunResult } from './types';
+import { runInKernel, DEFAULT_TIMEOUT_MS } from '../kernel/worker';
+import { summarize } from '../lib/summarize';
+import { setCellOutput, findCell } from './cells';
+import { current, setRunningCell } from './signals';
+import type { CellOutput, RunResult } from '../types';
 
 export interface RunSummary {
   cellId: string;
@@ -15,38 +17,6 @@ export interface RunSummary {
 const [lastRun, setLastRun] = createSignal<RunSummary | null>(null);
 const [timeoutMs, setTimeoutMs] = createSignal(DEFAULT_TIMEOUT_MS);
 export { lastRun, timeoutMs, setTimeoutMs };
-
-function describePart(p: OutputPart): string {
-  switch (p.kind) {
-    case 'table':
-      return 'table ' + (p.totalRows ?? 0) + ' rows x ' + (p.columns?.length ?? 0) + ' cols [' + (p.columns || []).slice(0, 8).join(', ') + ']';
-    case 'chart':
-      return 'chart(' + (p.spec?.type || '?') + ', ' + (p.spec?.data.datasets.length || 0) + ' series)';
-    case 'image':
-      return 'image (' + Math.round((p.src?.length || 0) / 1024) + ' KB)';
-    case 'json':
-      return 'json (' + (p.json?.length || 0) + ' chars)';
-    case 'markdown':
-      return 'markdown';
-    case 'error':
-      return 'error: ' + (p.message || '');
-    default: {
-      const t = (p.text || '').replace(/\s+/g, ' ').trim();
-      return t.length > 160 ? t.slice(0, 160) + '…' : t || 'empty';
-    }
-  }
-}
-
-/** One short line describing what a run produced — what the agent reads. */
-export function summarize(r: RunResult): string {
-  if (!r.ok) return (r.error?.name || 'Error') + ': ' + (r.error?.message || '');
-  const bits: string[] = [];
-  const logCount = (r.logs || []).length;
-  if (logCount) bits.push(logCount + ' log line' + (logCount === 1 ? '' : 's'));
-  for (const p of r.parts || []) bits.push(describePart(p));
-  if (r.savedTo) bits.push('saved to ' + r.savedTo);
-  return bits.length ? bits.join('; ') : 'no output';
-}
 
 function toOutput(r: RunResult): CellOutput {
   return {
