+++ apps/lab/src/styles/markdown.css
@@ -1,0 +1,62 @@
+/* Rendered markdown, for both markdown cells and md() output — lib/markdown.ts */
+
+.lab-md {
+  padding: var(--yaar-sp-2) var(--yaar-sp-3);
+  line-height: 1.6;
+}
+.lab-md-cell {
+  min-height: 24px;
+  cursor: text;
+  background: var(--yaar-bg);
+}
+.lab-md h1 {
+  font-size: var(--yaar-text-xl);
+  margin: 0.4em 0;
+}
+.lab-md h2 {
+  font-size: var(--yaar-text-lg);
+  margin: 0.4em 0;
+}
+.lab-md h3 {
+  font-size: var(--yaar-text-base);
+  margin: 0.4em 0;
+  color: var(--yaar-text-muted);
+}
+.lab-md p {
+  margin: 0.4em 0;
+}
+.lab-md code {
+  font-family: var(--yaar-font-mono);
+  font-size: var(--yaar-text-sm);
+  background: var(--yaar-bg-surface);
+  padding: 1px 4px;
+  border-radius: var(--yaar-radius-sm);
+}
+.lab-md pre {
+  background: var(--yaar-bg-surface);
+  padding: var(--yaar-sp-2);
+  border-radius: var(--yaar-radius-sm);
+  overflow: auto;
+}
+.lab-md pre code {
+  background: none;
+  padding: 0;
+}
+.lab-md a {
+  color: var(--yaar-accent);
+}
+.lab-md table {
+  border-collapse: collapse;
+  font-size: var(--yaar-text-sm);
+}
+.lab-md th,
+.lab-md td {
+  border: 1px solid var(--yaar-border);
+  padding: 2px 6px;
+}
+.lab-md blockquote {
+  border-left: 3px solid var(--yaar-border);
+  margin: 0.4em 0;
+  padding-left: var(--yaar-sp-3);
+  color: var(--yaar-text-muted);
+}
